Many improvements (see full commit message for info)
This commit adds the ability to create incidence objects with the
option for the user to give there own date grouping function. It
also adds a minimal incidence constructor and validator for
developers of other packages.

The commit also now allows list columns to be used. For this to
work we fall back on dplyr and data.frames for the aggregation.