provide a non data.table check for duplicate rows
Added a use of vctrs::vec_duplicate_any within the subsetting code for when
list columns are present.

We could only use the vctrs::vec_duplicate_any function but I want to
explicitly branch on non-data.table use as it will allow us to easily
drop dplyr support should we ever wish to do so.