Harsher and more verbose handling of POSIXct
See https://github.com/reconverse/incidence2/issues/104 for
motivation.

<POSIXct> objects are represented as seconds since the UNIX epoch
and, in our experience, this level of granularity is not normally
desired for aggregation.

This commit emphasise this to the user via a warning if present
in a call to `incidence()` and via an error if `complete_dates()`
is called on an <incidence2> object with a POSIXct date_index.
`complete_dates()` does provide an opt-in escape hatch via a new
argument (`allow_POSIXct`) should they still want the old
behaviour.

As part of this commit we also tweak other behaviour for consistency.
In particular we now allow users to specify `day` or `daily` as the
`interval` in a call to `incidence()`.  If these intervals are
specified we attempt to coerce the input to <Dates>. This conversion
is a thin wrapper around `as.Date()` that floors the underlying
data (to ensure whole numbers) and respects timezone of input.