Reintroduce plotting examples
Previously removed due to CRAN note that I could not reproduce:

```
Check: examples, Result: NOTE
  Examples with CPU (user + system) or elapsed time > 5s
                   user system elapsed
  plot.incidence2 7.479  0.192   0.913
  Examples with CPU time > 2.5 times elapsed time
                   user system elapsed ratio
  plot.incidence2 7.479  0.192   0.913 8.402
```

Getting to the bottom of it seems like it will be tricky as unsure if the
issue if in R/CRAN, data.table or my package. I wonder if there is a weird
interaction with the use of withAutoprint but not sure I want to spend much
time digging in to this.